+++ tasks/pipeline.py
@@ -3,7 +3,7 @@ from __future__ import print_function
 from invoke import task
 from invoke.exceptions import Exit
 
-from .deploy.trigger_agent_pipeline import trigger_agent_pipeline, wait_for_pipeline
+from .deploy.pipeline_tools import trigger_agent_pipeline, wait_for_pipeline
 from .deploy.gitlab import Gitlab
